Remove spin-locks in favor of multiple CrossProcessSemaphores.
Producer-consumer queues are often implemented with 2 semaphores that track how
much data/room is in the queue.  This is not that -- that makes the processes
more synchronized than is necessary _and_ requires us to be able to track how
much room is _left_, which qould require the semaphore to track the number of
bytes used/remaining, since we permit variable message lengths.  And
CrossProcessSemaphores can only increment/decrement one at a time (thanks,
Windows), which would make byte count tracking with semaphore operations O(n)
per transaction instead of O(1).

This implementation avoids all of that -- careful analysis bounds the semaphore
values to a maximum of 2.